+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
+### Changed
+
+## [2.7.0] - 2020-10-09
+
 ### Added
 
  - Action dct_check - Check if docker images are trusted
 
 ### Changed
+
   - Dependencies - Bump stripe to v2.54.0
   - Action usage report - Remove VPN from usage reporting
   - Action usage report - Report usage of nonfree deployments